+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Add a public api that allows third party apps to create activities
 
 # 2025.03.1